LangChain/CRAG(Cache RAG)/requirements.txt
faiss-cpu
numpy
sentence-transformers
langchain
PyPDF2
python-docx
groq